added functionality for login and create user, structure the main file in to management package
